% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{alignment}
\alias{alignment}
\title{Align soft clip reads}
\usage{
alignment(reference, fastq, bamOutput, threads = 1)
}
\arguments{
\item{reference}{The TE reference built with bowtie2}

\item{fastq}{The fastq file of soft clip reads, out from \bold{extractSoftClip}}

\item{bamOutput}{Character, name of output bam file}

\item{threads}{The number of cores for parallel computing.}
}
\description{
Align soft clip reads to the transposable elements reference.
}
