% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{buildIndex}
\alias{buildIndex}
\title{Build a Bowtie index from transposon sequences}
\usage{
buildIndex(reference, index, overwrite = FALSE, ...)
}
\arguments{
\item{reference}{Character, the path to the fasta reference containing the
transposon sequences.}

\item{index}{Character, the prefix of index.}

\item{overwrite}{Boolean, whether to overwrite existing files. Default is true.
default is 'TRUE'}

\item{...}{Other arguments}
}
\description{
Build a Bowtie index from transposon sequences
}
