% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{extractSoftClip}
\alias{extractSoftClip}
\title{Extract soft clip reads}
\usage{
extractSoftClip(file, outfq, mapq = 20, length = 30, tsd = 10)
}
\arguments{
\item{file}{Character, the alignment file. Supported formats include BAM and SAM.}

\item{outfq}{Character, the output FASTQ file.}

\item{mapq}{The MAPping quality, 10 by default.}

\item{length}{The minimum soft clip reads length, 30 by default.}

\item{tsd}{The max length of tsd.}
}
\description{
Extract soft clip reads from alignment.
}
