% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{processInsertion}
\alias{processInsertion}
\title{Detection of transposon insertion sites}
\usage{
processInsertion(file, max.gapwidth = 10)
}
\arguments{
\item{file}{InsertLocation function output}

\item{max.gapwidth}{The maximum distance between two site combined}
}
\description{
Transposon insertion site search and TSD sequenc assembly
}
